+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - The following server-side APIs to send data to the client:
   - `Connection::send_blocklist`
+  - `Connection::send_config_update_cmd`
   - `Connection::send_internal_network_list`
   - `Connection::send_reboot_cmd`
   - `Connection::send_tor_exit_node_list`